# proyecto-backend/app/Http/Resources/UserResource.php
<?php

namespace App\Http\Resources;

use Illuminate\Http\Request;
use Illuminate\Http\Resources\Json\JsonResource;

/**
 * @OA\Schema(
 *     title="UserResource",
 *     description="User resource",
 *     @OA\Xml(name="UserResource"),
 * )
 */
class UserResource extends JsonResource
{
    /**
     * Transform the resource into an array.
     *
     * @return array<string, mixed>
     *
     * @OA\Property(
     *     property="data",
     *     title="data",
     *     description="Data wrapper"
     * )
     *
     * @var \App\Models\User[]
     */
    public function toArray(Request $request): array
    {
        $roles = $this->roles;
        $news = $this->news;
        $comments = $this->comments;
        return [
            'id' => $this->id,
            'name' => $this->name,
            'email' => $this->email,
            'roles' => $roles,
            'news' => $news,
            'comments' => $comments,
            'created_at' => $this->created_at,
            'updated_at' => $this->updated_at,
        ];
    }
}