# proyecto-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Muestra un listado de usuarios",
                "description": "Devuelve un listado de usuarios",
                "operationId": "getUsersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "añade un usuario",
                "description": "devuelve los datos de un usuario",
                "operationId": "addUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/users/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "obtiene un usuario",
                "description": "devuelve un usuario",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id del usuario a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Edita un usuario",
                "description": "devuelve los datos de un usuario",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "ownUserOrAdmin": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Borra un usuario",
                "description": "Borra un usuario y devuelve un 204",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "admin": []
                    }
                ]
            }
        },
        "/api/assign_roles/{user}": {
            "put": {
                "tags": [
                    "AssignRoles"
                ],
                "summary": "asigna roles a un usuario",
                "description": "devuelve los roles de un usuario",
                "operationId": "updateAssignRoles",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Identificador del recurso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignRolesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    },
                    {
                        "sanctum": []
                    },
                    {
                        "admin": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "AssignRolesRequest": {
                "title": "Store User Request",
                "description": "Store AssignRoles request body data",
                "required": [
                    "roles"
                ],
                "properties": {
                    "roles": {
                        "title": "roles",
                        "description": "array con los roles del usuario",
                        "type": "string",
                        "example": [
                            "1",
                            "2"
                        ]
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AssignRolesRequest"
                }
            },
...
            "UserRequest": {
                "title": "Store User Request",
                "description": "Store User request body data",
                "required": [
                    "name,email,password"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "nombre del usuario",
                        "type": "string",
                        "example": "swaggerUser"
                    },
                    "email": {
                        "title": "email",
                        "description": "email del usuario",
                        "type": "string",
                        "example": "swaggerUser@email.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "password del usuario",
                        "type": "string",
                        "example": "12345678"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "confirmacion del password del usuario",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserRequest"
                }
            },
...
            "UserResource": {
                "title": "UserResource",
                "description": "User resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserResource"
                }
            },
...
            "User": {
                "properties": {
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string",
                        "example": "swaggerUser"
                    },
                    "email": {
                        "description": "email del usuario",
                        "type": "string",
                        "example": "swaggerUser@email.com"
                    },
                    "password": {
                        "description": "passord del usuario",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}